Added some apis and interfaces
