+++ firestore/crowdy-chat/models/campaigns.ts
@@ -1,0 +1,3 @@
+export interface ICampaigns {
+
+}
