+++ firestore/crowdy-chat/models/collections.ts
@@ -1,0 +1,3 @@
+export interface ICollection<T> {
+    [k: string]: T
+}
