+++ firestore/crowdy-chat/models/leads.ts
@@ -1,0 +1,3 @@
+export interface ILeads {
+
+}
