+++ firestore/crowdy-chat/models/members.ts
@@ -1,0 +1,3 @@
+export interface IMembers {
+
+}
