--- firestore/crowdy-chat/models/organizations.tsx
@@ -1,12 +1,0 @@
-export interface IOrganizations {
-    name: string
-
-    // Optional fields
-    image?: string
-    description?: string
-    industry?: string
-    website?: string
-
-    // This field will be used to store the openAI key
-    openAIKey?: string
-}
