+++ firestore/crowdy-chat/models/sources.ts
@@ -1,0 +1,3 @@
+export interface ISources {
+
+}
