+++ firestore/crowdy-chat/models/users.ts
@@ -1,0 +1,11 @@
+interface IUser {
+    email: string,
+    username?: string, // This is only needed if you are creating members
+    name?: string,
+    image?: string,
+
+    organizations: {
+        organizationId: string,
+        status: number
+    }[]
+}
