--- firestore/crowdy-chat/models/users.tsx
@@ -1,5 +1,0 @@
-interface IUser {
-    email: string,
-    name?: string,
-    image?: string
-}
